pyproject.toml
[project]
name = "nexus-agent"
version = "0.1.0"
description = "Nexus Agent - 新员工入职助手，基于LangChain的智能对话系统"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.0",
    "langchain-core>=1.2.5",
    "langchain-openai>=1.1.6",
    "langchain-community>=0.4.1",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pytest>=7.0.0",
    "rich>=13.0.0",
    "pydantic-settings>=2.0.0",
    # Sprint 2 RAG dependencies
    "langchain-text-splitters>=0.2.0",
    "pypdf>=3.0.0",
    "beautifulsoup4>=4.12.0",
    "langchain-chroma>=0.1.0",
    "sentence-transformers>=2.2.0",
    "langchain-huggingface>=0.1.0",
    # Sprint 4 Memory Management dependencies
    "redis>=5.0.0",
    "tiktoken>=0.5.0",
    # Sprint 5 API dependencies
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["nexus_agent/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''
